+++ zurch/__init__.py
@@ -1,6 +1,6 @@
 """zurch - A CLI search tool for Zotero installations."""
 
-__version__ = "0.6.2"
+__version__ = "0.6.4"
 
 from .cli import main
 from .search import ZoteroDatabase